@@ -12,8 +12,8 @@
   "dependencies": {
     "babel-polyfill": "^6.16.0",
     "ramda": "^0.22.1",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2",
     "react-event-listener": "^0.3.0",
     "react-hotkeys": "^0.9.0",
     "react-skylight": "^0.4.1",
