@@ -4,17 +4,16 @@
   "description": "",
   "main": "dist/app/main.js",
   "scripts": {
-    "start": "electron ./dist/app/main.js",
-    "build": "webpack --colors --config webpack/base.js && babel ./src/app/ -d ./dist/app/",
-    "lint": "eslint --ext .js --ext .jsx ./src ./test || true"
+    "start": "electron ./dist/app/main.js"
   },
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "babel-polyfill": "^6.16.0",
     "classnames": "^2.2.5",
     "ramda": "^0.22.1",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2",
     "react-event-listener": "^0.3.0",
     "react-hotkeys": "^0.9.0",
     "react-skylight": "^0.4.1",
@@ -26,47 +25,8 @@
     "xod-fs": "^0.0.1"
   },
   "devDependencies": {
-    "autoprefixer-loader": "^3.1.0",
-    "babel": "^5.8.23",
-    "babel-cli": "^6.4.0",
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-node6": "^11.0.0",
-    "babel-preset-react": "^6.3.13",
-    "babel-register": "^6.11.6",
-    "copy-webpack-plugin": "^3.0.1",
-    "css-loader": "^0.23.1",
     "devtron": "^1.3.0",
     "electron": "^1.4.1",
-    "eslint": "^2.13.0",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-import-resolver-webpack": "^0.3.2",
-    "eslint-plugin-import": "^1.8.1",
-    "eslint-plugin-jsx-a11y": "^1.5.3",
-    "eslint-plugin-react": "^5.2.2",
-    "exports-loader": "^0.6.3",
-    "file-loader": "^0.8.5",
-    "imports-loader": "^0.6.5",
-    "json5-loader": "^0.6.0",
-    "node-sass": "^3.8.0",
-    "npm-run-all": "^2.3.0",
-    "raw-loader": "^0.5.1",
-    "react-hot-loader": "^1.3.0",
-    "redux-devtools": "^3.3.1",
-    "redux-devtools-dispatch": "^2.2.0",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-filterable-log-monitor": "^0.6.6",
-    "redux-mock-store": "^1.1.1",
-    "sass-loader": "^4.0.0",
-    "source-map-support": "^0.4.2",
-    "style-loader": "^0.13.0",
-    "wallaby-webpack": "0.0.22",
-    "webpack": "^1.12.10",
-    "webpack-dev-server": "^1.14.0",
-    "webpack-merge": "^0.14.0",
-    "webpack-target-electron-renderer": "^0.4.0",
-    "webpack-validator": "^2.2.3"
+    "webpack-target-electron-renderer": "^0.4.0"
   }
 }
