@@ -1,40 +1,50 @@
+const fs = require('fs');
 const path = require('path');
 const webpack = require('webpack');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const webpackTargetElectronRenderer = require('webpack-target-electron-renderer');
 
+const pkgpath = subpath => path.resolve(__dirname, '..', subpath);
+const assetsPath = fs.realpathSync(pkgpath('node_modules/xod-client/src/core/assets'));
+
 const options = {
   devtool: 'source-map',
   entry: [
     'babel-polyfill',
-    './src/view/index.jsx',
+    pkgpath('src/view/index.jsx'),
+    pkgpath('node_modules/xod-client/src/core/styles/main.scss'),
+    pkgpath('src/view/styles/main.scss'),
   ],
   output: {
     filename: 'bundle.js',
-    path: path.join(__dirname, '../dist'),
+    path: pkgpath('dist'),
     publicPath: '',
   },
   devServer: {
     hot: true,
     host: 'localhost',
     port: 8080,
-    contentBase: './dist/',
+    contentBase: pkgpath('dist'),
   },
   resolve: {
-    root: path.join(__dirname, '../src'),
-    modulesDirectories: ['node_modules', 'src', 'src/node_modules'],
+    modulesDirectories: [
+      pkgpath('node_modules'),
+      pkgpath('node_modules/xod-client/node_modules'),
+      pkgpath('node_modules/xod-client/node_modules/xod-core/node_modules'),
+      pkgpath('node_modules/xod-espruino/node_modules'),
+    ],
     extensions: ['', '.js', '.jsx', '.scss'],
     alias: {
-      react: path.resolve('node_modules/react'),
       // @TODO: Get rid of this hack:
-      encoding: path.resolve('node_modules/react'),
-      'iconv-lite': path.resolve('node_modules/react'),
+      // encoding: pkgpath('node_modules/react'),
+      // 'iconv-lite': pkgpath('node_modules/react'),
     },
   },
   module: {
     loaders: [
       {
         test: /src\/.*\.jsx?$/,
+        exclude: /node_modules/,
         loaders: [
           'babel?presets[]=react,presets[]=es2015',
         ],
@@ -44,14 +54,13 @@ const options = {
         loaders: [
           'style',
           'css',
-          'autoprefixer?browsers=last 3 versions',
           'sass?outputStyle=expanded',
         ],
       },
       {
         test: /assets\/.*\.(jpe?g|png|gif|svg|ttf|eot|svg|woff|woff2)?$/,
         loaders: [
-          'file?name=assets/[path][name].[ext]?[hash:6]&context=../xod-client/src/core/assets',
+          `file?name=assets/[path][name].[ext]?[hash:6]&context=${assetsPath}`,
         ],
       },
       {
@@ -61,23 +70,23 @@ const options = {
         ],
       },
       {
-        test: /\.json5$/,
-        loader: 'json5',
+        test: /\.json$/,
+        loader: 'json-loader',
       },
       {
-        test: /src\/node_modules\/xod-espruino\/index\.js$/,
-        loader: 'null',
+        test: /\.json5$/,
+        loader: 'json5-loader',
       },
       {
         test: /json5\/lib\/require/,
         loader: 'null',
-      }
+      },
     ],
   },
   plugins: [
     new webpack.NoErrorsPlugin(),
     new CopyWebpackPlugin([
-      { from: 'src/view/index.html' },
+      { from: pkgpath('src/view/index.html') },
     ]),
   ],
 };
